app channel revision mostly adaptions that were brought over from the revised payment channel example to attain similarities between these examples

+++ app-channel/app/data.go
@@ -10,10 +10,10 @@ import (
 )
 
 // TicTacToeAppData is the app data struct.
-type TicTacToeAppData struct {
-	NextActor uint8
-	Grid      [9]FieldValue
-}
+type TicTacToeAppData struct { // Grid:
+	NextActor uint8         // 0 1 2
+	Grid      [9]FieldValue // 3 4 5
+} // 6 7 8
 
 func (d *TicTacToeAppData) String() string {
 	var b bytes.Buffer
@@ -42,7 +42,7 @@ func (d *TicTacToeAppData) Clone() channel.Data {
 }
 
 func (d *TicTacToeAppData) Set(x, y int, actorIdx channel.Index) {
-	if d.NextActor != uint8safe(actorIdx) {
+	if d.NextActor != uint8safe(uint16(actorIdx)) {
 		panic("invalid actor")
 	}
 	v := makeFieldValueFromPlayerIdx(actorIdx)
